requirements.txt
flask
flask_restplus
flask_api
redis
python-dotenv